+++ test/viking/model/class_properties/urlRoot.js
@@ -6,12 +6,38 @@
         equal(Model.urlRoot(), '/models');
     });
     
+    test("::urlRoot returns an URL based on #urlRoot set on the model", function () {
+        var Model = Viking.Model.extend('model', {
+            urlRoot: '/buoys'
+        });
+        equal(Model.urlRoot(), '/buoys');
+    });
+
     // STI test
     test("::urlRoot returns an URL based on modelName of the baseModel", function() {
         var Ship = Viking.Model.extend('ship');
         var Carrier = Ship.extend();
         
         equal(Carrier.urlRoot(), '/ships');
+    });
+
+    test("::urlRoot returns an URL based on #urlRoot set on the baseModel", function () {
+        var Ship = Viking.Model.extend('ship', {
+            urlRoot: '/myships'
+        });
+        var Carrier = Ship.extend();
+
+        equal(Carrier.urlRoot(), '/myships');
+    });
+
+
+    test("::urlRoot returns an URL based on #urlRoot set on the sti model", function () {
+        var Ship = Viking.Model.extend('ship');
+        var Carrier = Ship.extend({
+            urlRoot: '/carriers'
+        });
+
+        equal(Carrier.urlRoot(), '/carriers');
     });
 
 }());